Trait substrate_rpc::chain::ChainApi

pub trait ChainApi<Number, Hash, Header, SignedBlock>: Sized + Send + Sync + 'static {
    type Metadata: PubSubMetadata;
    fn header(&self, hash: Option<Hash>) -> Result<Option<Header>>;

    fn block(&self, hash: Option<Hash>) -> Result<Option<SignedBlock>>;

    fn block_hash(
        &self, 
        hash: Option<NumberOrHex<Number>>
    ) -> Result<Option<Hash>>;

    fn finalized_head(&self) -> Result<Hash>;

    fn subscribe_new_head(
        &self, 
        metadata: Self::Metadata, 
        subscriber: Subscriber<Header>
    );

    fn unsubscribe_new_head(
        &self, 
        metadata: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> RpcResult<bool>;

    fn subscribe_finalized_heads(
        &self, 
        metadata: Self::Metadata, 
        subscriber: Subscriber<Header>
    );

    fn unsubscribe_finalized_heads(
        &self, 
        metadata: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> RpcResult<bool>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>
    where
        Number: Send + Sync + 'static + DeserializeOwned,
        Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
        Header: Send + Sync + 'static + Serialize,
        SignedBlock: Send + Sync + 'static + Serialize,
    { ... }
}

Substrate blockchain API

Associated Types

type Metadata: PubSubMetadata

RPC metadata

Loading content...
Required methods

fn header(&self, hash: Option<Hash>) -> Result<Option<Header>>

Get header of a relay chain block.

fn block(&self, hash: Option<Hash>) -> Result<Option<SignedBlock>>

Get header and body of a relay chain block.

fn block_hash(&self, hash: Option<NumberOrHex<Number>>) -> Result<Option<Hash>>

Get hash of the n-th block in the canon chain.

By default returns latest block hash.

fn finalized_head(&self) -> Result<Hash>

Get hash of the last finalized block in the canon chain.

fn subscribe_new_head(
&self,
metadata: Self::Metadata,
subscriber: Subscriber<Header>
)

New head subscription

fn unsubscribe_new_head(
&self,
metadata: Option<Self::Metadata>,
id: SubscriptionId
) -> RpcResult<bool>

Unsubscribe from new head subscription.

fn subscribe_finalized_heads(
&self,
metadata: Self::Metadata,
subscriber: Subscriber<Header>
)

New head subscription

fn unsubscribe_finalized_heads(
&self,
metadata: Option<Self::Metadata>,
id: SubscriptionId
) -> RpcResult<bool>

Unsubscribe from new head subscription.

Loading content...
Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> where
Number: Send + Sync + 'static + DeserializeOwned,
Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
Header: Send + Sync + 'static + Serialize,
SignedBlock: Send + Sync + 'static + Serialize,

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...