Trait substrate_rpc::state::StateApi

pub trait StateApi<Hash>: 'static + Send + Sync {
    type Metadata: PubSubMetadata;
    fn call(
        &self, 
        name: String, 
        bytes: Bytes, 
        hash: Option<Hash>
    ) -> Result<Bytes, Error>;

    fn storage_keys(
        &self, 
        prefix: StorageKey, 
        hash: Option<Hash>
    ) -> Result<Vec<StorageKey>, Error>;

    fn storage(
        &self, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Result<Option<StorageData>, Error>;

    fn storage_hash(
        &self, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Result<Option<Hash>, Error>;

    fn storage_size(
        &self, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Result<Option<u64>, Error>;

    fn child_storage_keys(
        &self, 
        child_storage_key: StorageKey, 
        prefix: StorageKey, 
        hash: Option<Hash>
    ) -> Result<Vec<StorageKey>, Error>;

    fn child_storage(
        &self, 
        child_storage_key: StorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Result<Option<StorageData>, Error>;

    fn child_storage_hash(
        &self, 
        child_storage_key: StorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Result<Option<Hash>, Error>;

    fn child_storage_size(
        &self, 
        child_storage_key: StorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Result<Option<u64>, Error>;

    fn metadata(&self, hash: Option<Hash>) -> Result<Bytes, Error>;

    fn runtime_version(
        &self, 
        hash: Option<Hash>
    ) -> Result<RuntimeVersion, Error>;

    fn query_storage(
        &self, 
        keys: Vec<StorageKey>, 
        block: Hash, 
        hash: Option<Hash>
    ) -> Result<Vec<StorageChangeSet<Hash>>, Error>;

    fn subscribe_runtime_version(
        &self, 
        metadata: Self::Metadata, 
        subscriber: Subscriber<RuntimeVersion, Error>
    );

    fn unsubscribe_runtime_version(
        &self, 
        metadata: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> Result<bool, Error>;

    fn subscribe_storage(
        &self, 
        metadata: Self::Metadata, 
        subscriber: Subscriber<StorageChangeSet<Hash>, Error>, 
        keys: Option<Vec<StorageKey>>
    );

    fn unsubscribe_storage(
        &self, 
        metadata: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> Result<bool, Error>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>
    where
        Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
    { ... }
}

Substrate state API

Associated Types

type Metadata: PubSubMetadata

RPC Metadata

Loading content...
Required methods

fn call(
&self,
name: String,
bytes: Bytes,
hash: Option<Hash>
) -> Result<Bytes, Error>

Call a contract at a block's state.

fn storage_keys(
&self,
prefix: StorageKey,
hash: Option<Hash>
) -> Result<Vec<StorageKey>, Error>

Returns the keys with prefix, leave empty to get all the keys

fn storage(
&self,
key: StorageKey,
hash: Option<Hash>
) -> Result<Option<StorageData>, Error>

Returns a storage entry at a specific block's state.

fn storage_hash(
&self,
key: StorageKey,
hash: Option<Hash>
) -> Result<Option<Hash>, Error>

Returns the hash of a storage entry at a block's state.

fn storage_size(
&self,
key: StorageKey,
hash: Option<Hash>
) -> Result<Option<u64>, Error>

Returns the size of a storage entry at a block's state.

fn child_storage_keys(
&self,
child_storage_key: StorageKey,
prefix: StorageKey,
hash: Option<Hash>
) -> Result<Vec<StorageKey>, Error>

Returns the keys with prefix from a child storage, leave empty to get all the keys

fn child_storage(
&self,
child_storage_key: StorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Result<Option<StorageData>, Error>

Returns a child storage entry at a specific block's state.

fn child_storage_hash(
&self,
child_storage_key: StorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Result<Option<Hash>, Error>

Returns the hash of a child storage entry at a block's state.

fn child_storage_size(
&self,
child_storage_key: StorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Result<Option<u64>, Error>

Returns the size of a child storage entry at a block's state.

fn metadata(&self, hash: Option<Hash>) -> Result<Bytes, Error>

Returns the runtime metadata as an opaque blob.

fn runtime_version(&self, hash: Option<Hash>) -> Result<RuntimeVersion, Error>

Get the runtime version.

fn query_storage(
&self,
keys: Vec<StorageKey>,
block: Hash,
hash: Option<Hash>
) -> Result<Vec<StorageChangeSet<Hash>>, Error>

Query historical storage entries (by key) starting from a block given as the second parameter.

NOTE This first returned result contains the initial state of storage for all keys. Subsequent values in the vector represent changes to the previous state (diffs).

fn subscribe_runtime_version(
&self,
metadata: Self::Metadata,
subscriber: Subscriber<RuntimeVersion, Error>
)

New runtime version subscription

fn unsubscribe_runtime_version(
&self,
metadata: Option<Self::Metadata>,
id: SubscriptionId
) -> Result<bool, Error>

Unsubscribe from runtime version subscription

fn subscribe_storage(
&self,
metadata: Self::Metadata,
subscriber: Subscriber<StorageChangeSet<Hash>, Error>,
keys: Option<Vec<StorageKey>>
)

New storage subscription

fn unsubscribe_storage(
&self,
metadata: Option<Self::Metadata>,
id: SubscriptionId
) -> Result<bool, Error>

Unsubscribe from storage subscription

Loading content...
Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> where
Hash: Send + Sync + 'static + Serialize + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...