Trait substrate_rpc::state::StateBackend

pub trait StateBackend<B, E, Block: BlockT, RA>: Send + Sync + 'static where
    Block: BlockT<Hash = H256> + 'static,
    B: Backend<Block, Blake2Hasher> + Send + Sync + 'static,
    E: CallExecutor<Block, Blake2Hasher> + Send + Sync + 'static,
    RA: Send + Sync + 'static, {
    fn client(&self) -> &Arc<Client<B, E, Block, RA>>;

    fn subscriptions(&self) -> &Subscriptions;

    fn call(
        &self, 
        block: Option<Block::Hash>, 
        method: String, 
        call_data: Bytes
    ) -> FutureResult<Bytes>;

    fn storage_keys(
        &self, 
        block: Option<Block::Hash>, 
        prefix: StorageKey
    ) -> FutureResult<Vec<StorageKey>>;

    fn storage(
        &self, 
        block: Option<Block::Hash>, 
        key: StorageKey
    ) -> FutureResult<Option<StorageData>>;

    fn storage_hash(
        &self, 
        block: Option<Block::Hash>, 
        key: StorageKey
    ) -> FutureResult<Option<Block::Hash>>;

    fn child_storage_keys(
        &self, 
        block: Option<Block::Hash>, 
        child_storage_key: StorageKey, 
        prefix: StorageKey
    ) -> FutureResult<Vec<StorageKey>>;

    fn child_storage(
        &self, 
        block: Option<Block::Hash>, 
        child_storage_key: StorageKey, 
        key: StorageKey
    ) -> FutureResult<Option<StorageData>>;

    fn child_storage_hash(
        &self, 
        block: Option<Block::Hash>, 
        child_storage_key: StorageKey, 
        key: StorageKey
    ) -> FutureResult<Option<Block::Hash>>;

    fn metadata(&self, block: Option<Block::Hash>) -> FutureResult<Bytes>;

    fn runtime_version(
        &self, 
        block: Option<Block::Hash>
    ) -> FutureResult<RuntimeVersion>;

    fn query_storage(
        &self, 
        from: Block::Hash, 
        to: Option<Block::Hash>, 
        keys: Vec<StorageKey>
    ) -> FutureResult<Vec<StorageChangeSet<Block::Hash>>>;

    fn storage_size(
        &self, 
        block: Option<Block::Hash>, 
        key: StorageKey
    ) -> FutureResult<Option<u64>> { ... }

    fn child_storage_size(
        &self, 
        block: Option<Block::Hash>, 
        child_storage_key: StorageKey, 
        key: StorageKey
    ) -> FutureResult<Option<u64>> { ... }

    fn subscribe_runtime_version(
        &self, 
        _meta: Metadata, 
        subscriber: Subscriber<RuntimeVersion>
    ) { ... }

    fn unsubscribe_runtime_version(
        &self, 
        _meta: Option<Metadata>, 
        id: SubscriptionId
    ) -> RpcResult<bool> { ... }

    fn subscribe_storage(
        &self, 
        _meta: Metadata, 
        subscriber: Subscriber<StorageChangeSet<Block::Hash>>, 
        keys: Option<Vec<StorageKey>>
    ) { ... }

    fn unsubscribe_storage(
        &self, 
        _meta: Option<Metadata>, 
        id: SubscriptionId
    ) -> RpcResult<bool> { ... }
}

State backend API.

Required methods

fn client(&self) -> &Arc<Client<B, E, Block, RA>>

Get client reference.

fn subscriptions(&self) -> &Subscriptions

Get subscriptions reference.

fn call(
&self,
block: Option<Block::Hash>,
method: String,
call_data: Bytes
) -> FutureResult<Bytes>

Call runtime method at given block.

fn storage_keys(
&self,
block: Option<Block::Hash>,
prefix: StorageKey
) -> FutureResult<Vec<StorageKey>>

Returns the keys with prefix, leave empty to get all the keys.

fn storage(
&self,
block: Option<Block::Hash>,
key: StorageKey
) -> FutureResult<Option<StorageData>>

Returns a storage entry at a specific block's state.

fn storage_hash(
&self,
block: Option<Block::Hash>,
key: StorageKey
) -> FutureResult<Option<Block::Hash>>

Returns the hash of a storage entry at a block's state.

fn child_storage_keys(
&self,
block: Option<Block::Hash>,
child_storage_key: StorageKey,
prefix: StorageKey
) -> FutureResult<Vec<StorageKey>>

Returns the keys with prefix from a child storage, leave empty to get all the keys

fn child_storage(
&self,
block: Option<Block::Hash>,
child_storage_key: StorageKey,
key: StorageKey
) -> FutureResult<Option<StorageData>>

Returns a child storage entry at a specific block's state.

fn child_storage_hash(
&self,
block: Option<Block::Hash>,
child_storage_key: StorageKey,
key: StorageKey
) -> FutureResult<Option<Block::Hash>>

Returns the hash of a child storage entry at a block's state.

fn metadata(&self, block: Option<Block::Hash>) -> FutureResult<Bytes>

Returns the runtime metadata as an opaque blob.

fn runtime_version(
&self,
block: Option<Block::Hash>
) -> FutureResult<RuntimeVersion>

Get the runtime version.

fn query_storage(
&self,
from: Block::Hash,
to: Option<Block::Hash>,
keys: Vec<StorageKey>
) -> FutureResult<Vec<StorageChangeSet<Block::Hash>>>

Query historical storage entries (by key) starting from a block given as the second parameter.

NOTE This first returned result contains the initial state of storage for all keys. Subsequent values in the vector represent changes to the previous state (diffs).

Loading content...
Provided methods

fn storage_size(
&self,
block: Option<Block::Hash>,
key: StorageKey
) -> FutureResult<Option<u64>>

Returns the size of a storage entry at a block's state.

fn child_storage_size(
&self,
block: Option<Block::Hash>,
child_storage_key: StorageKey,
key: StorageKey
) -> FutureResult<Option<u64>>

Returns the size of a child storage entry at a block's state.

fn subscribe_runtime_version(
&self,
_meta: Metadata,
subscriber: Subscriber<RuntimeVersion>
)

New runtime version subscription

fn unsubscribe_runtime_version(
&self,
_meta: Option<Metadata>,
id: SubscriptionId
) -> RpcResult<bool>

Unsubscribe from runtime version subscription

fn subscribe_storage(
&self,
_meta: Metadata,
subscriber: Subscriber<StorageChangeSet<Block::Hash>>,
keys: Option<Vec<StorageKey>>
)

New storage subscription

fn unsubscribe_storage(
&self,
_meta: Option<Metadata>,
id: SubscriptionId
) -> RpcResult<bool>

Unsubscribe from storage subscription

Loading content...